Simplified tests.
 [x] Consolidated test_circ.py and test_rect.py
